# perf: use spread operator