--- a/index.js
+++ b/index.js
@@ -9,31 +9,15 @@ function createError (code, message, statusCode = 500, Base = Error) {
   code = code.toUpperCase()
   !statusCode && (statusCode = undefined)
 
-  function FastifyError (a, b, c) {
+  function FastifyError (...args) {
     if (!new.target) {
-      return new FastifyError(...arguments)
+      return new FastifyError(...args)
     }
     this.code = code
     this.name = 'FastifyError'
     this.statusCode = statusCode
 
-    // more performant than spread (...) operator
-    switch (arguments.length) {
-      case 3:
-        this.message = format(message, a, b, c)
-        break
-      case 2:
-        this.message = format(message, a, b)
-        break
-      case 1:
-        this.message = format(message, a)
-        break
-      case 0:
-        this.message = message
-        break
-      default:
-        this.message = format(message, ...arguments)
-    }
+    this.message = format(message, ...args)
 
     Error.stackTraceLimit !== 0 && Error.captureStackTrace(this, FastifyError)
   }